Revert "refactor(nomenclatura); Refatorado e corrigindo estrutura dos nomes e e organização da classe"
This reverts commit 1339c6b43d905e538816a33f7a6c73e95d5dc268.
