+++ desafio-casdastro/src/main/java/br/desafio/prodiga/BancoApi/ApiBancariaController.java
@@ -1,5 +1,5 @@
 
-package br.desafio.prodiga.bancoApi.controller;
+package br.desafio.prodiga.BancoApi;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
