+++ desafio-casdastro/src/main/java/br/desafio/prodiga/Enums/SituacaoFatura.java
@@ -1,4 +1,4 @@
-package br.desafio.prodiga.enums;
+package br.desafio.prodiga.Enums;
 
 public enum SituacaoFatura {
   GERADA, PAGA, CANCELADA
