+++ desafio-casdastro/src/main/java/br/desafio/prodiga/Repository/ClienteRepository.java
@@ -1,8 +1,8 @@
-package br.desafio.prodiga.repository;
+package br.desafio.prodiga.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import br.desafio.prodiga.model.Cliente;
+import br.desafio.prodiga.Model.Cliente;
 
 @Repository
 public interface ClienteRepository extends JpaRepository<Cliente, Long>{
